requirements.txt
fastapi
uvicorn
python-dotenv
scikit-learn
numpy
requests
google-genai
openai
